add cmake
