+++ .gitignore
@@ -1,3 +1,16 @@
 *.history*
 *bin*
 *.gch*
+.vscode/
+*.a
+CMakeLists.txt.user
+CMakeCache.txt
+CMakeFiles
+CMakeScripts
+Testing
+Makefile
+cmake_install.cmake
+install_manifest.txt
+compile_commands.json
+CTestTestfile.cmake
+_deps
