+++ README.md
@@ -1,5 +1,2 @@
 # f4
 Fast f4 algorithm to find groebner basis
-
-
-g++ main.cpp -o ../bin/test -O3 -mavx512f -mavx512cd -mavx512er -mavx512pf -mavx512dq -mavx512vl -mavx512ifma -mavx512vbmi -mavx512bw
