--- benchmarks/Makefile
@@ -1,11 +1,0 @@
-SUBDIRS = 	buchberger
-
-all: $(SUBDIRS)
-
-$(SUBDIRS):
-		$(MAKE) -C $@
-
-run: 
-	./bin/benchmark_buchberger
-
-.PHONY: all $(SUBDIRS) run
