+++ benchmarks/benchmarking.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <chrono>
 #include <iostream>
 #include <iomanip>
