+++ benchmarks/main.cpp
@@ -1,0 +1,6 @@
+#include "benchmarking.h"
+#include "buchberger.cpp"
+
+int main() {
+    benchmark_buchberger();
+}
