+++ lib/util/term.h
@@ -5,13 +5,15 @@
 
 namespace NUtils {
     class TTerm : public std::vector<uint64_t> {
+        private:
+            void Normalize();
         public:
             using TBase = std::vector<uint64_t>;
             using size_type = typename TBase::size_type;
 
             inline TTerm() : TBase() {}
 
-            inline explicit TTerm(size_type sz) : TBase(sz) {}
+            inline TTerm(size_type sz) : TBase(sz) {}
 
             inline TTerm(std::initializer_list<uint64_t> il) : TBase(il) {
                 Normalize();
@@ -27,7 +29,5 @@ namespace NUtils {
             friend TTerm lcm(const TTerm&, const TTerm&) noexcept;
 
             friend std::ostream& operator<<(std::ostream&, const TTerm&) noexcept;
-        private:
-            void Normalize();
     };
 }
